examples/telegram/requirements.txt
python-telegram-bot==21.10
anthropic>=0.40.0
python-dotenv>=1.0.0
mcp>=1.0.0
